+++ 3-Main-SoC-Realtek-RTL8196E/31-Bootloader/boot/include/tftp.h
@@ -13,6 +13,7 @@ extern void kick_tftpd(void);
 extern void tftpd_entry(void);
 extern unsigned long file_length_to_server;
 extern unsigned long image_address;
+extern unsigned long g_tftp_server_ip;
 extern void tftp_get_server_ip(unsigned char ip[4]);
 extern void tftp_set_server_ip(const unsigned char ip[4]);
 extern void tftp_set_server_mac(const unsigned char mac[6]);