+++ 3-Main-SoC-Realtek-RTL8196E/31-Bootloader/boot/include/ver.h
@@ -1,3 +1,3 @@
 // version header file
 
-static char B_VERSION[] = "V2.6";
+static char B_VERSION[] = "V2.7";